Setup possibility to build  nightly wheels on CI
This adds to possibility to build wheels by pushing to the `releases/nightly` branch which will trigger a new wheel build. These wheels won't be published on PyPI but will be available to download as release artefacts on the GitHub actions workflow. This is useful for testing changes to the wheel build scripts or to build a nightly version that might not be ready to be released as a stable version.