Standardise on 32-bit bitpacking. Closes #446.
Remove the option of uint64 bitpacking (previously the default) and
replace it with int32 bitpacking. Note the change to a signed type -
as TFL doesn't have an unsigned 32-bit type, and we don't use the
values as integers anyway so the sign doesn't matter, it's simpler
to always use int32 rather than uint32.

This is a large commit that touches almost every file in the repo;
the vast majority of the changes involve removing templating of the
bitpacking type. Otherwise, in the bconv kernel and bitpacking code,
functions needed to deal with non-int32 bitpacked data have been
stripped out.

The bitpacking type `TBitpacked` is defined in a single place
(types.h), and that definition is imported everywhere it's needed,
so changing the bitpacking type in the future should be
straightforward.