Improve performance with optimised Ruy matrix packing.
Implement a Ruy packing routine for 4x4 column-major kernel layouts.
This improves Arm32 and Aarch64 performance, as this is the kernel
layout that is used for the RHS (input activations) matrix in our
optimised assembly kernels, and the Ruy packing for this layout
currently falls back to a slow reference implementation.

Furthermore, enable caching of the packed LHS (weights) matrix.
Enable the `use_caching` flag within our op, and opt-in to caching
the LHS. The RHS doesn't benefit from caching as it contains input
activations which change for each inference.